% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_utilities.R
\name{generateEset}
\alias{generateEset}
\title{Generate Eset by number of mismatch and total normalize number}
\usage{
generateEset(m, n.mismatch = NULL, normalize = TRUE, normalize.total = 1e+06)
}
\arguments{
\item{m}{matrix}

\item{n.mismatch}{number of mismatch}

\item{normalize}{logical, default to TRUE}

\item{normalize.total}{default to 1e6}
}
\value{
an eset with
}
\description{
Generate Eset by number of mismatch and total normalize number
}
