% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_utilities.R
\name{getCountByMismatch}
\alias{getCountByMismatch}
\title{Get count matrix by number of mismatch}
\usage{
getCountByMismatch(
  count.dist,
  n.mismatch,
  normalize = TRUE,
  normalize.total = 1e+06,
  annotation = NULL,
  undertermined.col = NULL,
  ...
)
}
\arguments{
\item{count.dist}{count distribution matrix, originally outputed in count.dist.csv from ScreenBEAM.raw.count}

\item{n.mismatch}{numerical,number of mismatch}

\item{normalize}{logical}

\item{normalize.total}{total normalize number, default to 1e6}

\item{annotation}{NULL}

\item{undertermined.col}{NULL}

\item{...}{parameters pass to normalizeMiseqProfile}
}
\description{
Get count matrix by number of mismatch
}
