% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_utilities.R
\name{normalize.scale}
\alias{normalize.scale}
\title{Normalization function}
\usage{
normalize.scale(d, total = NULL, pseudoCount = 1)
}
\arguments{
\item{pseudoCount}{}
}
\description{
Normalization function
}
