% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_utilities.R
\name{normalizeMiseqProfile}
\alias{normalizeMiseqProfile}
\title{normalizeMiseqProfile}
\usage{
normalizeMiseqProfile(
  d,
  total = NULL,
  pseudoCount = 1,
  undertermined.col = "Undetermined"
)
}
\arguments{
\item{d}{data frame}

\item{total}{total normalize number}

\item{pseudoCount}{pseudo number to add, for eliminating zeros}

\item{undertermined.col}{default to 'Undetermined'}
}
\description{
normalizeMiseqProfile
}
