% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_utilities.R
\name{saveCountEset}
\alias{saveCountEset}
\title{saveCountEset: Save count data into different expression set with multiple number of mismatch value}
\usage{
saveCountEset(m, save.path, total = 1e+06, n.mismatch = 2)
}
\arguments{
\item{m}{matrix}

\item{save.path}{path to save your expression sets}

\item{total}{total normalization number}

\item{n.mismatch}{numeric, adjust according to mapping qc report}
}
\value{
Four expression sets
}
\description{
saveCountEset: Save count data into different expression set with multiple number of mismatch value
}
