% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_functions.R
\name{ScreenBEAM.check.lib}
\alias{ScreenBEAM.check.lib}
\title{Check sh/sgRNA library and Remove duplicated/substring sh/sgRNA}
\usage{
ScreenBEAM.check.lib(lib.path)
}
\arguments{
\item{lib.path}{character, the file path of the library. Should be a csv file. The 1st column is RNAid, 2nd column RNA sequence and 3rd column target gene name.}
}
\value{
Return a list object, containing whole library information after removed duplicates; The most command sh/sgRNA length; Total sh/sgRNA number; All the sh/sgRNA names;
Each sh/sgRNA length; Removed sh/sgRNA information.
}
\description{
\code{ScreenBEAM.check.lib} is a pre-processing function for \code{ScreenBEAM.raw.count}.
It requires the file path of the library file (.csv file, with 1st column to be RNAid, 2nd column RNA sequence, 3rd column target gene name).
It will remove the duplicated/substring sh/sgRNA and store it into a list.
}
