% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_utilities.R
\name{ScreenBEAM.createEset}
\alias{ScreenBEAM.createEset}
\title{ScreenBEAM.createEset: create eSet after mapping steps, including combining mapping info to your metadata and data normalization}
\usage{
ScreenBEAM.createEset(analysis.par, normalize.total = 1e+06, n.mismatch = 9)
}
\arguments{
\item{normalize.total, }{integer}

\item{n.mismatch, }{integer}
}
\description{
ScreenBEAM.createEset: create eSet after mapping steps, including combining mapping info to your metadata and data normalization
}
