% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_functions.R
\name{ScreenBEAM.dir.create}
\alias{ScreenBEAM.dir.create}
\title{ScreenBEAM.dir.create}
\usage{
ScreenBEAM.dir.create(
  project_main_dir = NULL,
  lib_name = NULL,
  DATE = TRUE,
  library_file = NULL,
  metadata_file = NULL,
  count_file = NULL,
  startFromFastq = TRUE
)
}
\arguments{
\item{project_main_dir}{character, name or absolute path of the main working directory.}

\item{lib_name}{character, name of the sh/sgRNA library. Each main working directory can contain multiple library sub-project folder.}

\item{DATE}{logical, if TRUE, current date information will be added to the sub-project folder name. Default is TRUE.}

\item{library_file}{character, path for the library file. Required when startFromFastq = TRUE.}

\item{metadata_file}{character, path for the metadata csv file. (sampleID,group,replicate,sampleLabel,fastqFile)}

\item{count_file}{character, path for the count tab file. (rnaID,geneID,Sample1,Sample2...). Required when startFromFastq = FALSE;}

\item{startFromFastq, }{logical, if TRUE, the project start from Fastq files.}
}
\value{
Return a list object, containing all the paths of the created folder.
}
\description{
\code{ScreenBEAM.dir.create} is used to help users create an organized working directory for the Functional Genomics Screening Projects.
It creates a hierarchcial working directory and returns a list contains this directory information.
}
\details{
This function needs users to define the main working directory and the library’s name. It creates
a main working directory with a subdirectory of the project(each library is a project). It also automatically creates 4
subfolders (fastq, metadata, library and output) within the project folder. fastq/, storing unzipped fastq files of each sample;
metadata/, storing the experiment design, describe detailed information of each sample; library/, storing sh/sgRNA library csv file and its fasta file (for detail, please check demo).
output/ folder contains 3 subfolder to store intermediate results generated by ScreenBEAM2 analysis. They are mapping/, QC/ and DR/.

Inside mapping/ there are 3 subfolders as well.
mapping/Step1_unique_fast, mapping/Step2_blat and mapping/Step3_raw_count. mapping/Step1_unique_fast stores unique fasta read from each sample's fastq file. mapping/Step2_blat stores
blat result when using unique fasta file and library fasta file. mapping/Step3_raw_count stores 3 raw count table created later in the pipeline.
This function also returns a list object (analysis.par in the demo) with directory information wrapped inside.
This list is an essential for ScreenBEAM2 analysis, all the important intermediate data generated later will be wrapped inside.

QC/ is used to store quality control analysis of the mapping result. Including library check and mapping statistics.
DR/ is used to store the Bayesian hierarchical meta analysis result.

If startFromFastq = FALSE, library_file is not required. Only metadata/, output/QC/ output/DR/ are created.
}
