% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_functions.R
\name{ScreenBEAM.gene.level}
\alias{ScreenBEAM.gene.level}
\title{GENE level meta-analysis of high-throughput Functional Genomics Screening analysis}
\usage{
ScreenBEAM.gene.level(
  input.file,
  control.samples,
  case.samples,
  control.groupname = "control",
  case.groupname = "treatment",
  gene.columnId = 2,
  data.type = c("microarray", "NGS"),
  do.normalization = TRUE,
  total = 1e+06,
  filterLowCount = FALSE,
  filterBy = "control",
  count.cutoff = 4,
  nitt = 15000,
  burnin = 5000,
  thin = 10,
  rna.size = 6,
  sample.rna.time = 100,
  method = "Bayesian",
  pooling = "partial",
  ...
)
}
\arguments{
\item{input.file, }{character, the file path of ".tsv" table with raw count number for each sample (column) and sh/sgRNA (row).}

\item{control.samples, }{vector of character, the control samples names, this information should match the column names of input.file. For example, c(T0_A, T0_B).}

\item{case.samples, }{vector of character, the case samples names, this information should match the column names of input.file. For example, c(T16_A, T16_B).}

\item{control.groupname, }{character, the name of your control samples' group name. For example, T0.}

\item{case.groupname, }{character, the name of your case samples' group name. For example, T16.}

\item{gene.columnId, }{integer, the number of the column which store the gene name. By default, 2.}

\item{data.type, }{character, can either be "microarray" or "NGS". Default is "microarray".}

\item{do.normalization, }{logical, if TRUE, a scaled normalization will be performed for each sample. To quantify this scale, user need to assign a value to \code{total}.
Default is TRUE.}

\item{total, }{integer, need to be larger than the colSums of the raw count table. Default is 1e6.}

\item{filterLowCount, }{logical, if TRUE, will remove rnas with low count, based on \code{filterBy} and \code{count.cutoff}. Default is TRUE.}

\item{filterBy, }{logical, if TRUE, will remove rnas with low count based on "control" samples or "case" samples. Default is "control".}

\item{count.cutoff, }{integer, the threshold of removing low count. Default is 4.}

\item{nitt, }{integer, the number of MCMC iterations. Default is 15000.}

\item{burnin, }{integer, burnin. Default is 5000.}

\item{thin, }{integer, thinning interval. Default is 10.}

\item{rna.size, }{integer, the most common number of sh/sgRNAs targetting the same gene. Default is 6.}

\item{sample.rna.time, }{integer, to remove unbalance number of sh/sgRNAs targetting the same gene. \code{rna.size} of sh/sgRNAs will be sampled, the sampling time.}

\item{method, }{character, estimation model. Either "Bayesian" or "MLE". Default is "Bayesian".}

\item{pooling, }{character, pooling method. Either "full" or "partial". Default is "partial".}

\item{...}{parameter pass to function DRAgeneLevel2}
}
\value{
Return a data frame containing all the gene-level meta-analysis statistical values. Including geneID, log2FC, z value, p value, FDR and so on.
}
\description{
\code{ScreenBEAM.gene.level} takes a ".tsv" table file with samples raw count number, user needs to assign the control sample's name and case sample's name vector.
It will do a gene-level meta-analysis of screening data. User also need to assign value to \code{rna.size} to the most common number of sh/sgRNAs targeting one gene.
This function will remove bia caused by unbalanced number of sh/sgRNAs targeting the same gene.
}
