% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_functions.R
\name{ScreenBEAM.mapping.QC}
\alias{ScreenBEAM.mapping.QC}
\title{Create an HTML report to perform quality control of library, mapping status}
\usage{
ScreenBEAM.mapping.QC(
  analysis.par,
  QC.Rmd.path = system.file("Rmd", "mapping_QC_report.Rmd", package = "ScreenBEAM2")
)
}
\arguments{
\item{analysis.par, }{the master list, containing unique reads files path, blat result files path.}

\item{QC.Rmd.path, }{character, the path where user put the Rmarkdown file of "mapping_QC_report". Default, it will call the one in ScreenBEAM2 package.}
}
\value{
It will create an Rmarkdown file and an HTML QC report in the folder of \code{analsis.par$out.dir.output.QC}.
}
\description{
\code{ScreenBEAM.mapping.QC} takes the master \code{analysis.par} list, which contains all the key data through ScreenBEAM meta-analysis.
It will call Rmarkdown file to create an HTML report of library quality control (show the deleted duplicated/substring sh/sgRNA table). Percentage of mapping.
Suggested \code{rna.size}, which is the most common number of sh/sgRNAs targeting the same gene. A barplot at the end shows the suggested number of mismatch
for user to rescure the most reads.
}
