% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_functions.R
\name{ScreenBEAM.Pairwise}
\alias{ScreenBEAM.Pairwise}
\title{GENE level meta-analysis of high-throughput Functional Genomics Screening analysis for pairwise comparison}
\usage{
ScreenBEAM.Pairwise(
  analysis.par,
  choose_level = "gene",
  use_index,
  case_group,
  control_group,
  gene.columnId = 2,
  data.type = c("microarray", "NGS"),
  do.normalization = TRUE,
  total = 1e+06,
  filterLowCount = TRUE,
  filterBy = "control",
  count.cutoff = 4,
  nitt = 15000,
  burnin = 5000,
  thin = 10,
  rna.size = 6,
  sample.rna.time = 100,
  method = "Bayesian",
  pooling = "partial",
  ...
)
}
\arguments{
\item{analysis.par, }{list}

\item{use_index, }{integer, the line number for the `analysis.par$norm.path` to analyze.}

\item{case_group, }{character, name for the case group}

\item{control_group, }{character, name for the control group}

\item{gene.columnId, }{integer, the number of the column which store the gene name. By default, 2.}

\item{data.type, }{character, can either be "microarray" or "NGS". Default is "microarray".}

\item{do.normalization, }{logical, if TRUE, a scaled normalization will be performed for each sample. To quantify this scale, user need to assign a value to \code{total}.
Default is TRUE.}

\item{total, }{integer, need to be larger than the colSums of the raw count table. Default is 1e6.}

\item{filterLowCount, }{logical, if TRUE, will remove rnas with low count, based on \code{filterBy} and \code{count.cutoff}. Default is TRUE.}

\item{filterBy, }{logical, if TRUE, will remove rnas with low count based on "control" samples or "case" samples. Default is "control".}

\item{count.cutoff, }{integer, the threshold of removing low count. Default is 4.}

\item{nitt, }{integer, the number of MCMC iterations. Default is 15000.}

\item{burnin, }{integer, burnin. Default is 5000.}

\item{thin, }{integer, thinning interval. Default is 10.}

\item{rna.size, }{integer, the most common number of sh/sgRNAs targetting the same gene. Default is 6.}

\item{sample.rna.time, }{integer, to remove unbalance number of sh/sgRNAs targetting the same gene. \code{rna.size} of sh/sgRNAs will be sampled, the sampling time.}

\item{method, }{character, estimation model. Either "Bayesian" or "MLE". Default is "Bayesian".}

\item{pooling, }{character, pooling method. Either "full" or "partial". Default is "partial".}

\item{...}{parameter pass to function ScreenBEAM.gene.level}
}
\value{
Return analysis.par
}
\description{
\code{ScreenBEAM.Pairwise} user needs to assign the control sample's name and case sample's name vector.
It will do a gene-level/rna-level meta-analysis of screening data. User also need to assign value to \code{rna.size} to the most common number of sh/sgRNAs targeting one gene.
This function will remove bia caused by unbalanced number of sh/sgRNAs targeting the same gene.
}
