% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenBEAM2_functions.R
\name{ScreenBEAM.trim.helper}
\alias{ScreenBEAM.trim.helper}
\title{Create a pdf plot to visualize the trim location.}
\usage{
ScreenBEAM.trim.helper(fastq.path, sample_num = 10000, pdf.file)
}
\arguments{
\item{fastq.path, }{character, the file path of one FASTQ reads file. Unzipped.}

\item{sample_num, }{integer, the random sample number of FASTQ reads when creating the barplot. Default is 10000.}

\item{pdf.file, }{character, the file path to store the output pdf plot.}
}
\value{
It will create a PDF barplot file showing A,T,C,G and N's percentage at each base position. User can pick the trim location based on the plot.
}
\description{
\code{ScreenBEAM.trim.helper} is a helper function for users who wants to know the most common trim place in their FASTQ reads.
This function requires the linux command \code{seqtk}. Please make sure you have it before calling this function.
}
